+++ src/system-test.js
@@ -999,7 +999,10 @@ export default class SystemTest {
    * @returns {Promise<void>}
    */
   async visit(path) {
-    await this.getCommunicator().sendCommand({type: "visit", path})
+    await timeout(
+      {timeout: this.getTimeouts(), errorMessage: `timeout while visiting path: ${path}`},
+      async () => await this.getCommunicator().sendCommand({type: "visit", path})
+    )
   }
 
   /**
@@ -1008,7 +1011,10 @@ export default class SystemTest {
    * @returns {Promise<void>}
    */
   async dismissTo(path) {
-    await this.getCommunicator().sendCommand({type: "dismissTo", path})
+    await timeout(
+      {timeout: this.getTimeouts(), errorMessage: `timeout while dismissing to path: ${path}`},
+      async () => await this.getCommunicator().sendCommand({type: "dismissTo", path})
+    )
   }
 
   /**